InputReader Functionality Changed Into Event Based System | Agent2DRenderer Implemented | Agent2DStateBase turned into MonoBehaviour from ScriptableObject

+++ Assets/Nojumpo/Scripts/Agent/2D/Movement Data/Agent2DMovementData.cs
@@ -6,7 +6,7 @@ namespace Nojumpo
     public class Agent2DMovementData : ScriptableObject
     {
         // -------------------------------- FIELDS --------------------------------
-        public float HorizontalMovementDirection;
+        public int HorizontalMovementDirection;
         public float CurrentSpeed;
         public Vector2 CurrentVelocity;
     }
